initial commit
This commit includes a Makefile with targets to build and install
software. It also contains a submodule for gst-rtsp-launch in
submodules/gst-rtsp-launch/; this program makes it trivial to capture
motion jpeg and stream it to any client that connects with rtsp.

The shell script src/pinecube-fixup.sh should be run once to reconfigure
the tmpfs for /run to work with systemd.

The shell script src/wireguard-setup.sh should be run once to enable
wireguard and to configure the interface. Peers need to be added
manually and this setup script is largely a convenience script to
automate the boring stuff.

The Makefile target `install-deps` installs many packages to allow for a
variety of possible streaming strategies. Some of the installed
packages are build dependencies and some are useful tools in their own
right.

Additional methods for streaming will be added to this repository at a
later date.